# src/main/resources/swagger-2.0.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8089","basePath":"/api","tags":[{"name":"authentication-controller","description":"Authentication Controller"},{"name":"gerenciador-viagem-controller","description":"Gerenciador Viagem Controller"},{"name":"status-aplicacao-controller","description":"Status Aplicacao Controller"}],"paths":{"/v1/auth":{"post":{"tags":["authentication-controller"],"summary":"Gera um Token de acesso","operationId":"gerarTokenJwtUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"authenticationDto","description":"authenticationDto","required":true,"schema":{"$ref":"#/definitions/JwtAuthenticationDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Response«TokenDto»"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/v1/status":{"get":{"tags":["status-aplicacao-controller"],"summary":"verificaStatusAplicacao","operationId":"verificaStatusAplicacaoUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/v1/viagem/{id}":{"delete":{"tags":["gerenciador-viagem-controller"],"summary":"Apaga uma viagem específica","operationId":"deleteUsingDELETE","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","description":"Authorization","required":true,"type":"string"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/v1/viagens":{"get":{"tags":["gerenciador-viagem-controller"],"summary":"Retorna todas as viagens","operationId":"listarUsingGET","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","description":"Authorization","required":true,"type":"string"},{"name":"regiao","in":"query","description":"regiao","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Response«List«Viagem»»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["gerenciador-viagem-controller"],"summary":"Cadastra uma viagem","operationId":"cadastrarUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","description":"Authorization","required":true,"type":"string"},{"in":"body","name":"viagemDto","description":"viagemDto","required":true,"schema":{"$ref":"#/definitions/ViagemDto"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/Response«Viagem»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/v1/viagens/{id}":{"get":{"tags":["gerenciador-viagem-controller"],"summary":"Retorna uma viagem específica","operationId":"buscarUsingGET","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","description":"Authorization","required":true,"type":"string"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Response«ViagemDtoResponse»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["gerenciador-viagem-controller"],"summary":"Atualiza uma viagem específica","operationId":"alterarUsingPUT","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","description":"Authorization","required":true,"type":"string"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"},{"in":"body","name":"viagemDto","description":"viagemDto","required":true,"schema":{"$ref":"#/definitions/ViagemDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Response«Viagem»"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"JwtAuthenticationDto":{"type":"object","properties":{"email":{"type":"string"},"senha":{"type":"string"}},"title":"JwtAuthenticationDto"},"Response«List«Viagem»»":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/definitions/Viagem"}},"errors":{"type":"array","items":{"type":"string"}}},"title":"Response«List«Viagem»»"},"Response«TokenDto»":{"type":"object","properties":{"data":{"$ref":"#/definitions/TokenDto"},"errors":{"type":"array","items":{"type":"string"}}},"title":"Response«TokenDto»"},"Response«ViagemDtoResponse»":{"type":"object","properties":{"data":{"$ref":"#/definitions/ViagemDtoResponse"},"errors":{"type":"array","items":{"type":"string"}}},"title":"Response«ViagemDtoResponse»"},"Response«Viagem»":{"type":"object","properties":{"data":{"$ref":"#/definitions/Viagem"},"errors":{"type":"array","items":{"type":"string"}}},"title":"Response«Viagem»"},"TokenDto":{"type":"object","properties":{"token":{"type":"string"}},"title":"TokenDto"},"Viagem":{"type":"object","properties":{"acompanhante":{"type":"string","description":"Nome do acompanhante da viagem"},"dataPartida":{"type":"string","description":"Data de partida da viagem (yyyy-MM-dd)"},"dataRetorno":{"type":"string","description":"Data de retorno da viagem (yyyy-MM-dd)"},"id":{"type":"integer","format":"int64","description":"Id da viagem"},"localDeDestino":{"type":"string","description":"Local de destino da viagem"},"regiao":{"type":"string","description":"Região de destino da viagem [Norte, Sul, Leste, Oeste]"}},"title":"Viagem"},"ViagemDto":{"type":"object","required":["localDeDestino"],"properties":{"acompanhante":{"type":"string","description":"Nome do acompanhante da viagem"},"dataPartida":{"type":"string","description":"Data de partida da viagem (yyyy-MM-dd)"},"dataRetorno":{"type":"string","description":"Data de retorno da viagem (yyyy-MM-dd)"},"localDeDestino":{"type":"string","description":"Local de destino da viagem"},"regiao":{"type":"string","description":"Região de destino da viagem","enum":["Norte","Sul","Leste","Oeste"]}},"title":"ViagemDto"},"ViagemDtoResponse":{"type":"object","properties":{"acompanhante":{"type":"string","description":"Nome do acompanhante da viagem"},"dataPartida":{"type":"string","description":"Data de partida da viagem"},"dataRetorno":{"type":"string","description":"Data de retorno da viagem"},"id":{"type":"integer","format":"int64","description":"Id da viagem"},"localDeDestino":{"type":"string"},"regiao":{"type":"string","description":"Região de destino da viagem [Norte, Sul, Leste, Oeste]"},"temperatura":{"type":"number","format":"float","description":"Temperatura prevista da região de destino"}},"title":"ViagemDtoResponse"}}}